requirements.txt
streamlit>=1.40.1
plotly
numpy
pandas
scipy
pyyaml
streamlit-keyup
